repair: extract repair_row
