+++ configure.py
@@ -856,6 +856,7 @@ scylla_core = (['database.cc',
                 'lister.cc',
                 'repair/hash.cc',
                 'repair/sink_source_for_repair.cc',
+                'repair/row.cc',
                 'repair/repair.cc',
                 'repair/row_level.cc',
                 'exceptions/exceptions.cc',
